feat: Phase 1-5 — DSP wiring, state recall, RT safety, polish
Architect plan: P0/P1/P2 all closed.

Phase 1 (P0 ship-blockers):
- Audition path: CriticalSection + MidiBuffer → SPSCQueue<RawMidiMsg, 256>
  with bounded drain ≤64/block, direct route to midiHandler_ (no
  MidiBuffer alloc on RT thread).
- releaseResources(): clears voice/filter/DC/envelopes/FX; also called
  from prepareToPlay before prepare() for SR-change correctness.
- Voice::prepare: hard-reset ampEnv/filterEnv/filter/dcBlocker.
- prepareToPlay primes smoothers (no first-block glide).
- setStateInformation reapply + reprime chain.
- Master gain double-application fixed (single path via smoother).

Phase 2 (state recall):
- APVTS expanded 8 → 70 params covering every PatchStruct field.
- buildPatchFromApvts / writePatchToApvts symmetric round-trip.
- AgentBridge consumer moved off audio thread to juce::Timer (20ms);
  pollPatch → writePatchToApvts → setValueNotifyingHost on message
  thread; engine reads via applyParameters every block.
- Typed parameter pointer cache (dynamic_cast + jassert at cache time).
- isBusesLayoutSupported (mono/stereo); getTailLengthSeconds = 5.0.
- pendingPrime_ atomic exchange in applyParameters (race-free).
- PatchStateManager dead code deleted; engine_juce collapsed into
  engine_core.

Phase 3 (P1 wiring + perf + smoothing):
- Wired previously-unwired DSP fields: osc[i] enabled/pan/volume/type,
  filter.type, voice_count cap, reverb.width (M/S blend),
  delay.bpm_sync, LFO targets Pan/WavetablePos/FmRatio.
- Per-osc rendering: OscSlot, Voice::renderStereo (per-osc mix → mono
  filter → per-osc-pan-weighted L/R split).
- Dual filter pre-alloc (MoogLadder + SVFilter); pointer swap on type
  change; no heap alloc on RT.
- Smoothers added: reverb width, filter drive (per voice), ADSR rates
  (amp + filter A/D/S/R, ~50ms tc), LFO depth+rate per LFO.
- Perf: cached portamentoAlpha_; std::pow(2, x/12) → exp2f; skip when
  semis==0.
- Delay: fractional interp (delaySamples_ float, linear i0/i1 blend);
  real ping-pong (L↔R cross-routed feedback at stereo=1).
- WavetableOscillator: mip formula one-octave-too-aggressive fix
  (ratio = N*phi, not N*phi*2); crossfade index resolution floor→ceil.
- FM oscillator: cached per-voice sample rate (was hardcoded 44100);
  independent carrier phase accumulator (was sharing modulator phase).
- Voice-pan single-stage (drop double-multiply with per-osc pan).

Phase 4 (P2 polish):
- Filter type-swap crossfade (~5ms blend old + new outputs).
- LFO amp clamp: voiceGain = max(0, 1 + lfoAmpMod) — no phase
  inversion when sum < -1.
- activeNotes(): documented RT-unsafe (no RT caller exists).
- ADSREnvelope release rescale on aborted attack with split coeff pairs
  (releaseCoeffStart_/releaseCoeff_) — survives applyPatch every block.
- MorphEngine loglerp ==0 guard (delay.time_s=0 bypass preserved
  through morph; asymmetric zero snaps to lo per documented convention).
- WebUiComponent: juce::Thread::launch → juce::ThreadPool with
  removeAllJobs(true, 5000ms) cancellation in dtor; bridge_ lifetime
  contract documented.
- Multi-frame default wavetable (sine → triangle → saw → square).
- Crossfade-during-crossfade snap-then-restart (no half-warmed outgoing).

Phase 5 (off-plan polish):
- LFO Pan/WavetablePos/FmRatio dropped slot-0 gate; all enabled oscs
  receive modulation (standard synth behavior).
- Direct unit tests for getTailLengthSeconds + isBusesLayoutSupported.

Tests: 319 → 382 passing. New files:
- tests/PluginLifecycleTest.cpp (~30 cases: SPSC, releaseResources,
  primeSmoothers, setState round-trip, APVTS coverage, AI pump, drain
  budget, audition end-to-end, SR-change silence, buses/tail).
- tests/UnwiredParamsTest.cpp (~10 cases: per-field wiring gates,
  LFO target wiring, LFO non-slot-0 modulation, WavetablePos shouldfail
  removed after multi-frame default landed).
- tests/Phase3FollowupTest.cpp (~10 cases: hard-pan direction,
  per-osc volume weighted mix, FM SR independence, filter swap mid-
  render, voice_count drop, BPM-sync numerical, LFO clamp folding,
  cross-cutting integration).
- tests/WebUiComponentWorkerPoolTest.cpp (worker pool lifecycle race).

Deleted: PatchStateManager.{h,cpp} + test (APVTS is now single source
of truth for patch state).

+++ src/engine/Delay.h
@@ -43,7 +43,7 @@ private:
     int bufferSize_{0};
 
     double sampleRate_{44100.0};
-    int delaySamples_{0};
+    float delaySamples_{0.0f};
     float feedback_{0.3f};
     float mix_{0.25f};
     float stereo_{0.5f};
